pyproject.toml
[tool.poetry]
authors = ["twitterdev team", "Chris Smith-Clarke <chrissmithlclarke@gmail.com>"]
description = "Fork of twitterdev/twauth-web using fastapi instead of flask, and poetry instead of pipenv"
license = "MIT"
name = "twauth-fastapi"
version = "0.1.0"

[tool.poetry.dependencies]
Jinja2 = "^3.0.1"
aiofiles = "^0.7.0"
fastapi = "^0.65.2"
python = "^3.8"
python-dotenv = "^0.18.0"
python-multipart = "^0.0.5"
requests = "^2.25.1"
requests-oauthlib = "^1.3.0"
uvicorn = "^0.14.0"

[tool.poetry.dev-dependencies]
black = "^21.6b0"
flake8 = "^3.9.2"
isort = "^5.9.1"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
